requirements.txt
datetime
robotframework
robotframework-faker
robotframework-appiumlibrary
robotframework-seleniumlibrary
robotframework-requests
robotframework-tidy
robotframework-jsonlibrary
python-dotenv
appium-python-client
robotframework-screencaplibrary
psycopg2-binary
pytz==2023.3
slacker
pyyaml
logging

